Added forecast api. Updated data and tests.

+++ app/src/main/java/me/maximpestryakov/yamblzweather/data/model/common/Main.java
@@ -9,10 +9,10 @@ public class Main {
     public float temp;
     @SerializedName("pressure")
     @Expose
-    public int pressure;
+    public float pressure;
     @SerializedName("humidity")
     @Expose
-    public int humidity;
+    public float humidity;
     @SerializedName("temp_min")
     @Expose
     public float tempMin;
